moving fieldToConceptIdMapping.js
